// src/models/user.ts
export interface User {
  id: number;
  name: string;
  email: string;
  created_at: string;
  updated_at: string;
  role: string;

  // constructor(id: number, name: string, email: string, created_at: string, updated_at: string, role: string){
  //   this.id =  id;
  //   this.name =  name;
  //   this.email =  email;
  //   this.created_at = created_at;
  //   this.updated_at = updated_at;
  //   this.role = role;
  // }
}